Initial changes for c++11 build on Linux/x11
In addition to adding '-std=c++11' to the CCFLAGS,
some macro usage had to be tweaking as to avoid
errors in patterns like X"...", where X would be a
string prefix in c++11. Luckily, whitespace solves
the problem.

Then there seems to be a problem because some embedded
libraries use C (as opposed to C++), but get passed the
same flags. I don't know how to tweak C-only flags, but
gcc just issues a warning that '-std=c++11' is simply
ignored. Whew!